requirements.txt
asttokens==2.2.1
backcall==0.2.0
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
executing==1.2.0
fonttools==4.39.2
ipython==8.11.0
jedi==0.18.2
kiwisolver==1.4.4
matplotlib==3.7.1
matplotlib-inline==0.1.6
mysql-connector-python==8.0.33
names==0.3.0
numpy==1.24.2
packaging==23.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.4.0
prompt-toolkit==3.0.38
pure-eval==0.2.2
Pygments==2.14.0
pyparsing==3.0.9
PyQt5==5.15.9
pyqt5-plugins==5.15.9.2.3
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
pyqt5-tools==5.15.9.3.3
python-dateutil==2.8.2
python-dotenv==1.0.0
qt5-applications==5.15.2.2.3
qt5-tools==5.15.2.1.3
six==1.16.0
stack-data==0.6.2
traitlets==5.9.0
wcwidth==0.2.6
